[Refactor][Norm] Say why the RMS norm dispatch functions sit outside the class
The op and kernel files each carry module-level functions next to a class, and
nothing in them told a reader which of those are forced. The custom_op pair is:
registration is import-time and once per qualified name, the schema is read off
the annotations so self cannot appear, and the instance comes back through a
string key. The boundary comment now states all three.

align_up and ALIGNMENT were forced by nothing, and nine modules defined them.
They move to a new kernels/tiling.py, taking the documented, guarded version
that reduction/_primitives.py had; the seven kernel files that copied the
unguarded one and the unused pair in ops/norm/norm_base.py are gone.
tileops.kernels.reduction.align_up still resolves.

RMSNormFwdOp now takes the normalized_shape product from norm_base's
normalized_shape_to_n instead of open-coding it, and stores _last_m rather than
(m, n) — n is self.N, settled at construction. Both match LayerNormFwdOp. Its
eps default is the signature default rather than a module constant; None still
normalizes to the same 1e-6, which the manifest allows.
