+++ src/tileops/ops/norm/norm_base.py
@@ -3,14 +3,7 @@
 import math
 from typing import Sequence
 
-__all__ = ["ALIGNMENT", "align_up", "normalized_shape_to_n"]
-
-ALIGNMENT = 256
-
-
-def align_up(n: int, alignment: int) -> int:
-    """Return ``n`` rounded up to the next multiple of ``alignment``."""
-    return ((n + alignment - 1) // alignment) * alignment
+__all__ = ["normalized_shape_to_n"]
 
 
 def normalized_shape_to_n(normalized_shape: Sequence[int]) -> int:
